fix: move stories to core/ directory
These styles are only part of the `core` brand and thus should only display in the `core` demos. Moving to this directory will show these only for core.
