+++ components/o-editorial-layout/stories/core/editorialLayout.scss
@@ -1,0 +1,3 @@
+@import "../../../../node_modules/@financial-times/o-editorial-layout/main";
+
+@include oEditorialLayout();
