--- components/o-editorial-layout/stories/editorialLayout.scss
@@ -1,3 +1,0 @@
-@import "@financial-times/o-editorial-layout/main";
-
-@include oEditorialLayout();
